ams-10812 reorganize api wrapper
